+++ src/pages/TimetablePage/index.tsx
@@ -7,15 +7,10 @@ import LectureList from 'components/TimetablePage/DefaultPage/LectureList';
 import MyLectureTimetable from 'components/TimetablePage/DefaultPage/MyLectureTimetable';
 import MyLectureList from 'components/TimetablePage/DefaultPage/MyLectureList';
 import Curriculum from 'components/TimetablePage/DefaultPage/Curriculum';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
-import {
-  myLectureAddLectureSelector, myLectureRemoveLectureSelector, myLecturesAtom, selectedSemesterAtom,
-} from 'utils/recoil/semester';
+import { useRecoilValue } from 'recoil';
+import { myLecturesAtom, selectedSemesterAtom } from 'utils/recoil/semester';
 import useTokenState from 'utils/hooks/useTokenState';
 import useTimetableInfoList from 'components/TimetablePage/hooks/useTimetableInfoList';
-import useAddTimetableLecture from 'components/TimetablePage/hooks/useAddTimetableLecture';
-import { LectureInfo, TimetableLectureInfo } from 'interfaces/Lecture';
-import useDeleteTimetableLecture from 'components/TimetablePage/hooks/useDeleteTimetableLecture';
 import styles from './TimetablePage.module.scss';
 
 function TimetablePage() {
@@ -29,31 +24,6 @@ function TimetablePage() {
 
   const myLectures = token ? (myLecturesFromServer ?? []) : (myLecturesFromLocalStorageValue ?? []);
 
-  const { mutate: mutateAddWithServer } = useAddTimetableLecture(token);
-  const addLectureToLocalStorage = useSetRecoilState(myLectureAddLectureSelector);
-
-  const addMyLecture = (clickedLecture: LectureInfo) => {
-    if (token) {
-      mutateAddWithServer({
-        semester: selectedSemester,
-        timetable: [{ class_title: clickedLecture.name, ...clickedLecture }],
-      });
-    } else {
-      addLectureToLocalStorage(clickedLecture);
-    }
-  };
-
-  const { mutate: removeLectureFromServer } = useDeleteTimetableLecture(selectedSemester, token);
-  const removeLectureFromLocalStorage = useSetRecoilState(myLectureRemoveLectureSelector);
-
-  const removeMyLecture = (clickedLecture: LectureInfo | TimetableLectureInfo) => {
-    if ('name' in clickedLecture) {
-      removeLectureFromLocalStorage(clickedLecture);
-      return;
-    }
-    removeLectureFromServer(clickedLecture.id.toString());
-  };
-
   return (
     <div className={styles.page}>
       {!isMobile ? (
@@ -61,11 +31,11 @@ function TimetablePage() {
           <h1 className={styles.page__title}>시간표</h1>
           <div className={styles.page__content}>
             {/* 강의 목록 */}
-            <LectureList myLectures={myLectures} addMyLecture={addMyLecture} />
+            <LectureList myLectures={myLectures} />
             {/* 나의 시간표 타임 테이블 */}
             <MyLectureTimetable myLectures={myLectures} />
             {/* 나의 시간표 강의 목록 */}
-            <MyLectureList myLectures={myLectures} removeMyLectures={removeMyLecture} />
+            <MyLectureList myLectures={myLectures} />
             {/* 시간표 커리큘럼 */}
             <Curriculum />
           </div>
